refactor: 변경된 타입 및 패키지 적용
